Move IbexUtil.collectExtensions to editor utils project; use in both GT and TGG editor
