+++ org.emoflon.ibex.tgg.editor/META-INF/MANIFEST.MF
@@ -82,5 +82,4 @@ Export-Package: antlr,
  org.moflon.tgg.mosl.tgg,
  org.moflon.tgg.mosl.tgg.impl,
  org.moflon.tgg.mosl.tgg.util,
- org.moflon.tgg.mosl.validation,
- org.moflon.util
+ org.moflon.tgg.mosl.validation
