+++ org.emoflon.ibex.tgg.ide/META-INF/MANIFEST.MF
@@ -23,6 +23,7 @@ Require-Bundle: com.google.guava,
  org.eclipse.xtext,
  org.eclipse.xtext.ui,
  org.eclipse.xtext.xbase.lib,
+ org.emoflon.ibex.common.editor.utils,
  org.emoflon.ibex.tgg.editor,
  org.moflon.core.plugins,
  org.moflon.core.ui,
