+++ org.emoflon.ibex.tgg.ide/src/org/emoflon/ibex/tgg/ide/admin/IbexTGGBuilder.java
@@ -46,6 +46,7 @@ import org.eclipse.emf.ecore.xmi.impl.URIHandlerImpl;
 import org.eclipse.xtext.EcoreUtil2;
 import org.eclipse.xtext.resource.XtextResource;
 import org.eclipse.xtext.resource.XtextResourceSet;
+import org.emoflon.ibex.common.editor.utils.ExtensionsUtil;
 import org.emoflon.ibex.tgg.ide.transformation.EditorTGGtoFlattenedTGG;
 import org.moflon.core.utilities.LogUtils;
 import org.moflon.core.utilities.WorkspaceHelper;
@@ -56,7 +57,6 @@ import org.moflon.tgg.mosl.tgg.ComplementRule;
 import org.moflon.tgg.mosl.tgg.Nac;
 import org.moflon.tgg.mosl.tgg.Rule;
 import org.moflon.tgg.mosl.tgg.TripleGraphGrammarFile;
-import org.moflon.util.IbexUtil;
 
 public class IbexTGGBuilder extends IncrementalProjectBuilder implements IResourceDeltaVisitor {
 	public static final String INTERNAL_TGG_MODEL_EXTENSION = ".tgg.xmi";
@@ -78,7 +78,7 @@ public class IbexTGGBuilder extends IncrementalProjectBuilder implements IResour
 	private Map<String, Object> blackboard;
 
 	public IbexTGGBuilder() {
-		builderExtensions = IbexUtil.collectExtensions(IBUILDER_EXTENSON_ID, "class", BuilderExtension.class);
+		builderExtensions = ExtensionsUtil.collectExtensions(IBUILDER_EXTENSON_ID, "class", BuilderExtension.class);
 	}
 	
 	@Override
